test: Adding new types to collection tests and simplifying scalar tests
